added select1 test with PDO update

+++ libsnowflakeclient/examples/select1.c
@@ -25,7 +25,7 @@ int main() {
     /* query */
     SNOWFLAKE_STMT *sfstmt = snowflake_stmt(sf);
     SNOWFLAKE_BIND_OUTPUT c1;
-    int out = 0;
+    int64 out = 0;
     c1.idx = 1;
     c1.type = SF_C_TYPE_INT64;
     c1.value = (void *) &out;
